Feat: Comment 도메인 관련 예외 클래스 추가

+++ backend/src/main/java/kr/dgucaps/capsv4/domain/comment/service/CommentService.java
@@ -1,14 +1,18 @@
 package kr.dgucaps.capsv4.domain.comment.service;
 
+import kr.dgucaps.capsv4.domain.board.exception.BoardNotFoundException;
 import kr.dgucaps.capsv4.domain.comment.dto.CreateCommentRequest;
 import kr.dgucaps.capsv4.domain.board.entity.Board;
+import kr.dgucaps.capsv4.domain.comment.entity.Comment;
+import kr.dgucaps.capsv4.domain.comment.exception.CommentNotAuthorException;
+import kr.dgucaps.capsv4.domain.comment.exception.CommentNotFoundException;
 import kr.dgucaps.capsv4.domain.comment.repository.CommentRepository;
 import kr.dgucaps.capsv4.domain.user.entity.User;
 import kr.dgucaps.capsv4.domain.board.repository.BoardRepository;
+import kr.dgucaps.capsv4.domain.user.exception.UserNotFoundException;
 import kr.dgucaps.capsv4.domain.user.repository.UserRepository;
 import kr.dgucaps.capsv4.security.SecurityUtil;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,15 +28,21 @@ public class CommentService {
     @Transactional
     public void createComment(CreateCommentRequest request) {
         User user = userRepository.findByUserId(SecurityUtil.getCurrentUserName())
-                .orElseThrow(() -> new UsernameNotFoundException("해당 회원을 찾을 수 없습니다."));
+                .orElseThrow(() -> new UserNotFoundException(SecurityUtil.getCurrentUserName()));
         Board board = boardRepository.findById(request.getBoardId())
-                .orElseThrow(() -> new IllegalArgumentException("해당 게시글을 찾을 수 없습니다."));
+                .orElseThrow(() -> new BoardNotFoundException(request.getBoardId()));
         commentRepository.save(request.toEntity(board, user));
     }
 
     @Transactional
     public void deleteComment(Integer commentId) {
-        //TODO: 예외처리
+        User user = userRepository.findByUserId(SecurityUtil.getCurrentUserName())
+                .orElseThrow(() -> new UserNotFoundException(SecurityUtil.getCurrentUserName()));
+        Comment comment = commentRepository.findById(commentId)
+                .orElseThrow(() -> new CommentNotFoundException(commentId));
+        if (!user.getUserId().equals(comment.getUser().getUserId())) {
+            throw new CommentNotAuthorException();
+        }
         commentRepository.deleteById(commentId);
     }
 }
